feat: add animation for settings tabs content changing
